done sort

+++ app/src/main/java/com/waka/dana/na/domain/model/StorageItem.kt
@@ -8,6 +8,6 @@ data class StorageItem(
     val name: String?,
     val path: String?,
     val lastModified: Long?,
-    val size: Double?,
+    val size: Long?,
     val isFolder: Boolean?
 )
